修复判空处理，代码优化

+++ snackbar/src/main/java/com/chenyacheng/snackbar/SnackBarManager.java
@@ -88,7 +88,7 @@ final class SnackBarManager {
                 durationMs = LONG_DURATION_MS;
             }
             mHandler.removeCallbacksAndMessages(r);
-            mHandler.sendMessageDelayed(Message.obtain(mHandler, MSG_TIMEOUT, r), (long) durationMs);
+            mHandler.sendMessageDelayed(Message.obtain(mHandler, MSG_TIMEOUT, r), durationMs);
         }
     }
